+++ tests/RoutingHelpersTest.php
@@ -34,4 +34,60 @@ class RoutingHelpersTest extends BrowserKitTestCase
 
         $this->assertEquals($expected_url, $url);
     }
+
+    public function test_document_download_url_generation()
+    {
+        $user = $this->createAdminUser();
+        $document = $this->createDocument($user);
+
+        $url = RoutingHelpers::download($document);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/download", $url);
+        
+        $url_with_version = RoutingHelpers::download($document, $document->file);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/download/{$document->file->uuid}", $url_with_version);
+    }
+    
+    public function test_document_embed_url_generation()
+    {
+        $user = $this->createAdminUser();
+        $document = $this->createDocument($user);
+
+        $url = RoutingHelpers::embed($document);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/download?embed=true", $url);
+        
+        $url_with_version = RoutingHelpers::embed($document, $document->file);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/download/{$document->file->uuid}?embed=true", $url_with_version);
+    }
+
+    public function test_document_preview_url_generation()
+    {
+        $user = $this->createAdminUser();
+        $document = $this->createDocument($user);
+
+        $url = RoutingHelpers::preview($document);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/preview", $url);
+        
+        $url_with_version = RoutingHelpers::preview($document, $document->file);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/preview/{$document->file->uuid}", $url_with_version);
+    }
+
+    public function test_document_thumbnail_url_generation()
+    {
+        $user = $this->createAdminUser();
+        $document = $this->createDocument($user);
+
+        $url = RoutingHelpers::thumbnail($document);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/thumbnail", $url);
+        
+        $url_with_version = RoutingHelpers::thumbnail($document, $document->file);
+
+        $this->assertStringEndsWith("/klink/$document->local_document_id/thumbnail/{$document->file->uuid}", $url_with_version);
+    }
 }